user-service/requirements.txt
asgiref==3.11.0
cfgv==3.5.0
distlib==0.4.0
django==5.2.9
django-filter==25.2
django-redis==6.0.0
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
filelock==3.20.1
gunicorn==23.0.0
identify==2.6.15
nodeenv==1.10.0
packaging==25.0
platformdirs==4.5.1
pre-commit==4.5.1
psycopg2==2.9.11
pyjwt==2.10.1
pyyaml==6.0.3
redis==7.1.0
ruff==0.14.10
sqlparse==0.5.5
virtualenv==20.35.4
